requirements.txt
python-dotenv
discord.py
selenium
python-dateutil
